[QNS] Fix to reporting cellular in rat preference 3 when cellular becomes available.
Fixed an issue in rat preference (wifi when no Cellular) not
reporting cellular even though there is cellular available.

Rat preference (wifi when no cellular) should be reported cellular
when cellular is available. However, if the handover is disallowed
while the PDN is connected, the opposite transport type should be
reported to disconnect PDN over iwlan and then move to cellular
properly.

Bug: 231965279
Test: atest QualifiedNetworksServiceTests:AccessNetworkEvaluatorTest
Change-Id: I0e951e012f47e5b57e299fd8b44e43a373f1a8e9