[QNS] Evaluate one more time when iwlan is available for SIM hotswap
 QNS reported an empty list as a qulified network for the hotswap case.
However, If the wifi keeps as connected constantly during sim hot swap,
iwlan could not be reported because there was no event to evaluate.
If Wi-Fi stays connected during SIM hot swap and there is no changes
at cellular side, Qns is not able to report iwlan as a qualified network
because there are no events to evaluate.
 Evaluating one more time when iwlan is available for ANE rebuild, ANE
will have a chance to report iwlan.

Bug: 242302712
Test: QualifiedNetworksServiceTests
Change-Id: If230c5bac76fc55ee913a7fd1ce3b527f98e8a14